+++ samples/semantickernel-concepts/semantickernel-syntax-examples/src/main/java/com/microsoft/semantickernel/samples/syntaxexamples/memory/JDBC_DataStorage.java
@@ -1,28 +1,26 @@
-// Copyright (c) Microsoft. All rights reserved.
 package com.microsoft.semantickernel.samples.syntaxexamples.memory;
 
 import com.azure.ai.openai.OpenAIAsyncClient;
 import com.azure.ai.openai.OpenAIClientBuilder;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.credential.KeyCredential;
-import com.azure.core.util.ClientOptions;
-import com.azure.core.util.MetricsOptions;
-import com.azure.core.util.TracingOptions;
-import com.azure.search.documents.indexes.SearchIndexAsyncClient;
-import com.azure.search.documents.indexes.SearchIndexClientBuilder;
 import com.microsoft.semantickernel.aiservices.openai.textembedding.OpenAITextEmbeddingGenerationService;
-import com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStore;
-import com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStoreOptions;
-import com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStoreRecordCollection;
-import com.microsoft.semantickernel.data.VolatileVectorStore;
-import com.microsoft.semantickernel.data.VolatileVectorStoreRecordCollection;
+import com.microsoft.semantickernel.connectors.data.jdbc.JDBCVectorStore;
+import com.microsoft.semantickernel.connectors.data.jdbc.JDBCVectorStoreOptions;
+import com.microsoft.semantickernel.connectors.data.jdbc.MySQLVectorStoreQueryProvider;
+import com.microsoft.semantickernel.data.VectorStoreRecordCollection;
 import com.microsoft.semantickernel.data.recordattributes.VectorStoreRecordDataAttribute;
 import com.microsoft.semantickernel.data.recordattributes.VectorStoreRecordKeyAttribute;
 import com.microsoft.semantickernel.data.recordattributes.VectorStoreRecordVectorAttribute;
+import com.mysql.cj.jdbc.MysqlDataSource;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import javax.sql.DataSource;
 import java.nio.charset.StandardCharsets;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.Base64;
 import java.util.Collections;
@@ -30,17 +28,19 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-public class InMemory_DataStorage {
+public class JDBC_DataStorage {
+
     private static final String CLIENT_KEY = System.getenv("CLIENT_KEY");
     private static final String AZURE_CLIENT_KEY = System.getenv("AZURE_CLIENT_KEY");
 
     // Only required if AZURE_CLIENT_KEY is set
     private static final String CLIENT_ENDPOINT = System.getenv("CLIENT_ENDPOINT");
-
-    // Embedding model configuration
     private static final String MODEL_ID = System.getenv()
-        .getOrDefault("EMBEDDING_MODEL_ID", "text-embedding-3-large");
+            .getOrDefault("EMBEDDING_MODEL_ID", "text-embedding-3-large");
     private static final int EMBEDDING_DIMENSIONS = 1536;
+
+    // Run a MySQL server with:
+    // docker run -d --name mysql-container -e MYSQL_ROOT_PASSWORD=root -e MYSQL_DATABASE=sk -p 3306:3306 mysql:latest
 
     static class GitHubFile {
         @VectorStoreRecordKeyAttribute()
@@ -52,11 +52,15 @@ public class InMemory_DataStorage {
         @VectorStoreRecordVectorAttribute(dimensions = EMBEDDING_DIMENSIONS, indexKind = "Hnsw")
         private final List<Float> embedding;
 
+        public GitHubFile() {
+            this(null, null, null, Collections.emptyList());
+        }
+
         public GitHubFile(
-            String id,
-            String description,
-            String link,
-            List<Float> embedding) {
+                String id,
+                String description,
+                String link,
+                List<Float> embedding) {
             this.id = id;
             this.description = description;
             this.link = link;
@@ -77,75 +81,92 @@ public class InMemory_DataStorage {
         }
     }
 
-    public static void main(String[] args) {
-        System.out.println("===================================================================");
-        System.out.println("========== Volatile (In memory) Vector Store Example ==============");
-        System.out.println("===================================================================");
+    public static void main(String[] args) throws SQLException {
+        System.out.println("==============================================================");
+        System.out.println("========== JDBC Vector Store Example ==============");
+        System.out.println("==============================================================");
 
         OpenAIAsyncClient client;
 
         if (AZURE_CLIENT_KEY != null) {
             client = new OpenAIClientBuilder()
-                .credential(new AzureKeyCredential(AZURE_CLIENT_KEY))
-                .endpoint(CLIENT_ENDPOINT)
-                .buildAsyncClient();
+                    .credential(new AzureKeyCredential(AZURE_CLIENT_KEY))
+                    .endpoint(CLIENT_ENDPOINT)
+                    .buildAsyncClient();
 
         } else {
             client = new OpenAIClientBuilder()
-                .credential(new KeyCredential(CLIENT_KEY))
-                .buildAsyncClient();
+                    .credential(new KeyCredential(CLIENT_KEY))
+                    .buildAsyncClient();
         }
 
         var embeddingGeneration = OpenAITextEmbeddingGenerationService.builder()
-            .withOpenAIAsyncClient(client)
-            .withModelId(MODEL_ID)
-            .withDimensions(EMBEDDING_DIMENSIONS)
-            .build();
+                .withOpenAIAsyncClient(client)
+                .withModelId(MODEL_ID)
+                .withDimensions(EMBEDDING_DIMENSIONS)
+                .build();
 
-        inMemoryDataStorage(embeddingGeneration);
+        var dataSource = new MysqlDataSource();
+        dataSource.setUrl("jdbc:mysql://localhost:3306/sk");
+        dataSource.setPassword("root");
+        dataSource.setUser("root");
+
+        dataStorageWithMySQL(dataSource, embeddingGeneration);
     }
 
-    public static void inMemoryDataStorage(
-        OpenAITextEmbeddingGenerationService embeddingGeneration) {
-        // Create a new Volatile vector store
-        var volatileVectorStore = new VolatileVectorStore();
+    public static void dataStorageWithMySQL(
+            DataSource dataSource,
+            OpenAITextEmbeddingGenerationService embeddingGeneration) {
+
+        // Build a query provider
+        var queryProvider = MySQLVectorStoreQueryProvider.builder()
+                .withDataSource(dataSource)
+                .build();
+
+        // Create a new vector store
+        var jdbcVectorStore = JDBCVectorStore.builder()
+                .withDataSource(dataSource)
+                .withOptions(JDBCVectorStoreOptions.builder()
+                        .withQueryProvider(queryProvider)
+                        .build())
+                .build();
 
         String collectionName = "skgithubfiles";
-        var collection = volatileVectorStore.getCollection(collectionName, GitHubFile.class, null);
+        var collection = jdbcVectorStore.getCollection(collectionName, GitHubFile.class,
+                null);
 
         // Create collection if it does not exist and store data
         List<String> ids = collection
-            .createCollectionIfNotExistsAsync()
-            .then(storeData(collection, embeddingGeneration, sampleData()))
-            .block();
+                .createCollectionIfNotExistsAsync()
+                .then(storeData(collection, embeddingGeneration, sampleData()))
+                .block();
 
-        // Retrieve all records from the collection
         List<GitHubFile> data = collection.getBatchAsync(ids, null).block();
 
         data.forEach(gitHubFile -> System.out.println("Retrieved: " + gitHubFile.getDescription()));
     }
 
     private static Mono<List<String>> storeData(
-        VolatileVectorStoreRecordCollection<GitHubFile> recordCollection,
-        OpenAITextEmbeddingGenerationService embeddingGeneration,
-        Map<String, String> data) {
+            VectorStoreRecordCollection<String, GitHubFile> recordStore,
+            OpenAITextEmbeddingGenerationService embeddingGeneration,
+            Map<String, String> data) {
 
         return Flux.fromIterable(data.entrySet())
-            .flatMap(entry -> {
-                System.out.println("Save '" + entry.getKey() + "' to memory.");
+                .flatMap(entry -> {
+                    System.out.println("Save '" + entry.getKey() + "' to memory.");
 
-                return embeddingGeneration
-                    .generateEmbeddingsAsync(Collections.singletonList(entry.getValue()))
-                    .flatMap(embeddings -> {
-                        GitHubFile gitHubFile = new GitHubFile(
-                            GitHubFile.encodeId(entry.getKey()),
-                            entry.getValue(),
-                            entry.getKey(),
-                            embeddings.get(0).getVector());
-                        return recordCollection.upsertAsync(gitHubFile, null);
-                    });
-            })
-            .collectList();
+                    return embeddingGeneration
+                            .generateEmbeddingsAsync(Collections.singletonList(entry.getValue()))
+                            .flatMap(embeddings -> {
+                                GitHubFile gitHubFile = new GitHubFile(
+                                        GitHubFile.encodeId(entry.getKey()),
+                                        entry.getValue(),
+                                        entry.getKey(),
+                                        embeddings.get(0).getVector());
+                                return recordStore.upsertAsync(gitHubFile, null);
+                            });
+                })
+                .collectList();
     }
 
     private static Map<String, String> sampleData() {