+++ semantickernel-experimental/src/main/java/com/microsoft/semantickernel/connectors/data/jdbc/JDBCVectorStoreRecordCollectionFactory.java
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft. All rights reserved.
+package com.microsoft.semantickernel.connectors.data.jdbc;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+
+/**
+ * Factory for creating JDBC vector store record collections.
+ */
+public interface JDBCVectorStoreRecordCollectionFactory {
+    /**
+     * Creates a new JDBC vector store record collection.
+     *
+     * @param options The options for the collection.
+     * @return The new JDBC vector store record collection.
+     */
+    <Record> JDBCVectorStoreRecordCollection<Record> createVectorStoreRecordCollection(
+        DataSource dataSource,
+        String collectionName,
+        JDBCVectorStoreRecordCollectionOptions<Record> options);
+}