+++ semantickernel-experimental/src/main/java/com/microsoft/semantickernel/data/VectorStoreRecordCollection.java
@@ -9,41 +9,40 @@ import reactor.core.publisher.Mono;
 import java.util.List;
 
 public interface VectorStoreRecordCollection<Key, Record> {
-
     /**
      * Gets the name of the collection.
      *
      * @return The name of the collection.
      */
-    public String getCollectionName();
+    String getCollectionName();
 
     /**
      * Checks if the collection exists in the store.
      *
      * @return A Mono emitting a boolean indicating if the collection exists.
      */
-    public Mono<Boolean> collectionExistsAsync();
+    Mono<Boolean> collectionExistsAsync();
 
     /**
      * Creates the collection in the store.
      *
      * @return A Mono representing the completion of the creation operation.
      */
-    public Mono<Void> createCollectionAsync();
+    Mono<Void> createCollectionAsync();
 
     /**
      * Creates the collection in the store if it does not exist.
      *
      * @return A Mono representing the completion of the creation operation.
      */
-    public Mono<Void> createCollectionIfNotExistsAsync();
+    Mono<Void> createCollectionIfNotExistsAsync();
 
     /**
      * Deletes the collection from the store.
      *
      * @return A Mono representing the completion of the deletion operation.
      */
-    public Mono<Void> deleteCollectionAsync();
+    Mono<Void> deleteCollectionAsync();
 
     /**
      * Gets a record from the store.